Organize all the re-exports into a single file
Hypothetically, Core.swift could be copy-pasted unmodified into any renderer
