+++ Sources/TokamakDOM/Views/Picker.swift
@@ -15,13 +15,6 @@
 import JavaScriptKit
 import TokamakCore
 
-public typealias Picker = TokamakCore.Picker
-public typealias PopUpButtonPickerStyle = TokamakCore.PopUpButtonPickerStyle
-public typealias RadioGroupPickerStyle = TokamakCore.RadioGroupPickerStyle
-public typealias SegmentedPickerStyle = TokamakCore.SegmentedPickerStyle
-public typealias WheelPickerStyle = TokamakCore.WheelPickerStyle
-public typealias DefaultPickerStyle = TokamakCore.DefaultPickerStyle
-
 extension _PickerContainer: ViewDeferredToRenderer {
   public var deferredBody: AnyView {
     AnyView(HTML("label") {
